+++ src/imio/email/parser/__init__.py
@@ -1,0 +1,3 @@
+from email.policy import default
+
+email_policy = default